# GitMiner/Documentation/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /gitminer/users/{id}:
    get:
      tags:
      - User
      - find
      - get
      - one
      - user
      - users
      summary: Retrieve a user by ID.
      description: Retrieve a user based on the ID.
      operationId: findOne
      parameters:
      - name: id
        in: path
        description: ID of user to be retrieved.
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags:
      - User
      - update
      - user
      - users
      - put
      summary: Updates a user.
      description: Updates a user.
      operationId: update
      parameters:
      - name: id
        in: path
        description: Id of user to be updated.
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "204":
          description: No Content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          description: Bad Request
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
    delete:
      tags:
      - User
      - destroy
      - user
      - delete
      - users
      summary: Deletes a user.
      description: Deletes a user.
      operationId: delete
      parameters:
      - name: id
        in: path
        description: Id of user to deleted.
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          description: Bad Request
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
...
  /gitminer/users:
    get:
      tags:
      - all
      - User
      - find
      - get
      - user
      - users
      summary: Retrieve a list of all users.
      description: Retrieve a list of all users.
      operationId: findALl
      parameters:
      - name: page
        in: query
        description: Page to retrieve.
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: activeusers
        in: query
        description: Order by activity.
        required: false
        schema:
          type: boolean
      - name: order
        in: query
        required: false
        schema:
          type: string
      - name: size
        in: query
        description: Size of the content of the page.
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
    post:
      tags:
      - User
      - post
      - create
      - user
      - users
      summary: Create a user.
      description: Create a user.
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /gitminer/user/{name}:
    get:
      tags:
      - User
      - find
      - get
      - one
      - name
      - user
      - users
      summary: Retrieve a user by name.
      description: Retrieve a user based on the name.
      operationId: findOneByName
      parameters:
      - name: name
        in: path
        description: Name of user to be retrieved.
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
...
components:
  schemas:
    User:
      required:
      - username
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        name:
          type: string
        avatar_url:
          type: string
        web_url:
          type: string
...